+++ translations/location_history_visualizer_de.ts
@@ -180,6 +180,18 @@
         <translation>Hell</translation>
     </message>
     <message>
+        <source>Midnight</source>
+        <translation>Mitternacht</translation>
+    </message>
+    <message>
+        <source>Nord</source>
+        <translation>Nord</translation>
+    </message>
+    <message>
+        <source>Sepia</source>
+        <translation>Sepia</translation>
+    </message>
+    <message>
         <source>Zoom in</source>
         <translation>Hineinzoomen</translation>
     </message>