+++ translations/location_history_visualizer_id.ts
@@ -180,6 +180,18 @@
         <translation>Terang</translation>
     </message>
     <message>
+        <source>Midnight</source>
+        <translation>Tengah malam</translation>
+    </message>
+    <message>
+        <source>Nord</source>
+        <translation>Nord</translation>
+    </message>
+    <message>
+        <source>Sepia</source>
+        <translation>Sepia</translation>
+    </message>
+    <message>
         <source>Zoom in</source>
         <translation>Perbesar</translation>
     </message>