+++ translations/location_history_visualizer_vi.ts
@@ -180,6 +180,18 @@
         <translation>Sáng</translation>
     </message>
     <message>
+        <source>Midnight</source>
+        <translation>Nửa đêm</translation>
+    </message>
+    <message>
+        <source>Nord</source>
+        <translation>Nord</translation>
+    </message>
+    <message>
+        <source>Sepia</source>
+        <translation>Sepia</translation>
+    </message>
+    <message>
         <source>Zoom in</source>
         <translation>Phóng to</translation>
     </message>